REPAUTO-40: move statuses select into settings modal
