+++ src/frontend/components/Settings/Settings.jsx
@@ -1,22 +1,24 @@
 import React from "react";
 
-import { Textfield } from '@forge/react';
+import { Textfield, Label, Select } from '@forge/react';
 import BaseModal from "../BaseModal/BaseModal";
-import useSettings from "./useSettings";
 
 const Settings = ({ 
     isVisible, 
     setIsVisible,    
     slackEndpoint,
     setSlackEndpoint,
-    setSetting 
+    setSetting,
+    statuses,
+    handleSelectChange,
+    selectedStatuses
 }) => {
-
     return (
         <BaseModal 
             isVisible={isVisible} 
             setIsVisible={setIsVisible} 
             title="Settings"
+            sx={{ height: 700 }}
         >
             <Textfield 
                 width={600} 
@@ -27,6 +29,18 @@ const Settings = ({
                 name='Slack Channel Endpoint' 
                 placeholder='Slack Endpoint' 
                 value={slackEndpoint}
+            />
+            <Label labelFor="multi-select-example">
+                Select statuses to include in summary
+            </Label>
+            <Select
+                inputId="multi-select-example"
+                options={statuses}
+                defaultValue={selectedStatuses}
+                isMulti
+                isSearchable={false}
+                placeholder="Choose a status"
+                onChange={(value) => handleSelectChange(value)}
             />
         </BaseModal>        
     );
